LayoutTests/imported/w3c/web-platform-tests/html/interaction/focus/focusgroup/tentative/grid-navigation/simple-case-with-non-focusable-cell-in-the-center-expected.txt

r1c1	r1c2	r1c3
r2c1	r2c2	r2c3
r3c1	r3c2	r3c3

FAIL A right arrow press should move the focus to the next column, skipping the non-focusable cell. assert_equals: expected Element node <td id="r2c3" tabindex="-1">r2c3</td> but got Element node <td id="r2c1" tabindex="-1">r2c1</td>
FAIL A down arrow press should move the focus to the next row, skipping the non-focusable cell. assert_equals: expected Element node <td id="r3c2" tabindex="-1">r3c2</td> but got Element node <td id="r1c2" tabindex="-1">r1c2</td>
FAIL A left arrow press should move to the previous column, skipping the non-focusable cell. assert_equals: expected Element node <td id="r2c1" tabindex="-1">r2c1</td> but got Element node <td id="r2c3" tabindex="-1">r2c3</td>
FAIL An up arrow press should move the focus to the previous row, skipping the non-focusable cell. assert_equals: expected Element node <td id="r1c2" tabindex="-1">r1c2</td> but got Element node <td id="r3c2" tabindex="-1">r3c2</td>
